requirements.txt
autopep8==2.0.0
black==22.12.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pylibdmtx==0.1.10
PyMuPDF==1.21.1
python-dotenv==0.21.0
tomli==2.0.1
treepoem==3.17.0
Werkzeug==2.2.2
zxing==0.14
